rewrite `contexts/contexts_bi` using a zipper
Rewrite `contexts` and `contexts_bi` using a Zipper, improving
performance by 37% in the `contexts enum big` benchmark.

DETAILS

My annoyances with the previous implementation were the following:

- It recursively chained iterators together. Chained iterators are slow.

- The closures recursively called themselves, and copied the entire
  model at each level.

With the zipper, the implementation is simpler, the iterator no longer
has any recursion or chaining, and the returned context closures have
been simplified. Instead of the closures calling nesting closures for
each level in the tree, the closure holds the zipper and traverses up it
instead. I have also tried to reduce the number of clones in general.

The zipper implementation here would be a good basis for a proper Zipper
interface using Uniplate.