feat: add specialization helpers for derive
Add specialisation helpers for the derive macro, that return a default
implementation for `Biplate<To>` if the input does not implement
`Biplate<To>`.

The eventual aim of this change is to remove the `walk_into` attribute,
and allow the derive macro to automatically infer which types we can
walk into when deriving `Biplate<T>`. This commit just includes the
specialization helpers needed to make this work: changing the derive
macro will be done in a future commit.

DETAILS

The `try_biplate_to!` macro is the front-end of this specialisation
code. `try_biplate_to!(x,T)` tries to call `Biplate<T>` on `x` -  if `x`
implements `Biplate<T>`, it returns `x.biplate()`, otherwise it returns
a default implementation of `Biplate<T>` that returns a `Tree::Zero`,
denoting that the `x` contains no children of type `T`.

This uses the auto-deref specialization trick
(<https://lukaskalbertodt.github.io/2019/12/05/generalized-autoref-based-specialization.html>)
to provide two implementations of the `spez_try_biplate` method. To do
this, I define the following traits, both containing `spez_try_biplate` as a
method:

1. `BiplateYes`, applicable when `Self: Biplate<T>`. Its
   implementation of `spez_try_biplate` calls `Biplate<T>::biplate` on
   the argument passed to it.

2. `BiplateNo`, applicable always. Its implementation of
   `spez_try_biplate` returns the default implementation of
   `Biplate<T>`,

`BiplateYes` has a higher priority than `BiplateNo`, making `BiplateNo`
act like a fallback implementation when the type does not implement
`Biplate<T>`.

To make this trick work:

* The traits are implemented on a wrapper type `SpezBiplate`, not
  the actual value types.
* Both traits must be in scope.
* The method must be called by doing `((&&SpezBiplate::(x,PhantomData::<T>)).spez_try_biplate()`.

The `try_biplate_to!` macro encapsulates this boilerplate to make using
this easier.

In the upcoming derive macro changes, we will call `try_biplate_to!` on
every field, and if that field does not implement `Biplate`,
`try_biplate_to!` will give a default implementation. This removes the
need to manually specify which types to walk into (i.e. call
`Biplate<T>` recursively on) for each biplate instance.

BACKGROUND

`Biplate<T>` looks inside a value for all children of type `T`. The key
feature of `Biplate` is that it can traverse through nested types when
looking for a value. For example, if a value of type `Expresssion`
contains a `Stmt` that contains a `String`, `Biplate<String> for
Expression` should return that string.

Currently, we require the user to explictly state which types to walk
into for each `Biplate` instance. In the below example, we explicitly
tell `Biplate` to look inside `Expr` for strings, so `Biplate<String>
for Stmt` returns both the strings inside `Expr::Var` and
  `Stmt::Assign`.

```rs
enum Stmt {
    Assign(String, Expr),
    Sequence(Vec<Stmt>),
    If(Expr, Box<Stmt>, Box<Stmt>),
    While(Expr, Box<Stmt>),
}

enum Expr {
    Add(Box<Expr>, Box<Expr>),
    Sub(Box<Expr>, Box<Expr>),
    Mul(Box<Expr>, Box<Expr>),
    Div(Box<Expr>, Box<Expr>),
    Val(i32),
    Var(String),
    Neg(Box<Expr>),
}
```
If we called `Biplate<String>` for all types inside `Expr`, instead of
just those in the `walk_into` list, Rust would error, telling us that
i32 does not implement `Biplate<String>`. Furthermore, as Rust does not
support trait specialisation (multiple impl blocks applying to the same
trait), specifying a default implementation of `Biplate<String>` is not
possible.

This commit gets around this language limitation by using the
autoderef-specializaton trick
(<https://lukaskalbertodt.github.io/2019/12/05/generalized-autoref-based-specialization.html>),
which gives us specialisation-like behaviour in stable Rust. This allows
us to specify a default implementation for Biplate, and thus
automatically infer which types to walk into when deriving biplate.