refactor(derive): simplify types
Simplify the type data structure and its parsing. We no longer use the
concept wrapper types or Type::Unplateable atall (this was a failed
early experiment in detecting which instances to derive), so this commit
removes these.

All we need to know is whether a type is boxed or not. Not having
wrapper types significantly simplifies type parsing, paving the way for
supporting other kind of types, such as tuple types.